requirements.txt
# Core Django framework
Django==5.2.7

# Django REST Framework for API functionality
djangorestframework==3.16.1

# CORS headers for React Native/frontend integration
django-cors-headers==4.9.0

# HTTP requests library (for testing)
requests==2.32.5

# Database drivers (uncomment if needed)
# psycopg2-binary==2.9.9  # For PostgreSQL
# mysqlclient==2.2.0      # For MySQL

# Development and deployment
gunicorn==21.2.0          # WSGI server for production
python-decouple==3.8      # Environment variables management
whitenoise==6.6.0         # Static files serving

# Optional utilities
Pillow==10.1.0            # Image processing (if handling images)
python-dateutil==2.8.2   # Date utilities